backend/requirements.txt
fastapi==0.112.2
openai==1.42.0
oracledb==2.4.1
requests==2.32.3
PyPDF2==3.0.1
python-magic==0.4.27
docx2txt==0.8
chardet==5.2.0
python-docx==1.1.2
flask-talisman==1.1.0
Flask-WTF==1.2.1
Flask-Limiter==3.8.0
python-dotenv==1.0.1
